Updates to MyAPS2 and prepro_scadsandspline
Should run  as normal without any intervention. Made a few changes to make it friendlier for single trial processing, as well as issues that would make it difficult to run on other machines or for other studies. This includes: change getcon_MyAPS2_ERP so that an second argument will make it return single trial condition numbers. Relatedly, prepro_scadsandspline now can take multiple arguments in cell array instead of just log file. Prepro_scads can now also take 'all' for the conditions to find and it will take all unique strings found by the getcon function. EEGlab will break if a condition is not present because it was artifact rejected, so it now checks to make sure that it exists before. Baseline period is no long hardcoded, but will revert to [-100 0] if not specified (I think baselining that earlier in the script may cause problems though). The new baseline argument can also be set to 'none' for no baselining. When using SaveAvgFile in prepro, the trigger and sample rate are no longer hardcoded.